Add in bugfix for sockscan2 issue.
Sockscan2 on certain images would create a buffer address space of a
particular size, then try to locate objects within it.  If the object
started late on in the buffer, the scanner would try and access objects
beyond the end of the buffer.  Read would return '' rather than None,
and cause an Exception.

This initial fix ensures that obj reads always return None if the
address space return either None or ''.